pyproject.toml
[tool.poetry]
name = "podaj-lapke-api"
version = "0.1.0"
description = ""
authors = ["Filip <mularskif@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
gunicorn = "^20.1.0"
whitenoise = "^6.4.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"
psycopg2 = "^2.9.6"
django-environ = "^0.10.0"
drf-spectacular = "^0.26.2"
django-cors-headers = "^4.0.0"
pillow = "^9.5.0"
geopy = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
